Use cyw43_wifi_get_rssi() for rssi updates (new in SDK 1.5.0).
- The return type is int32_t.

- In poll mode, cyw43_wifi_get_rssi() may not be called in an ISR,
  and must be called from the core on which networking was initialized.

The last condition means that rssi updates must be implemented
differently for poll mode. In background mode, we run the updates
from the repeating timer on core1, as before. In poll mode, the
repeating timer just sets a boolean to indicate the timeout, upon
which the update is run from the main loop on core0.